POLIO-1574: refactor subactivities round dates
- Change layout to mirror rounds tab
    - Move create button in RoundDates component
    - rename RoundDates -> SubActivitiesInfos
    - Simplify Grid structure and re-align with table
