+++ plugins/polio/js/src/domains/Campaigns/SubActivities/SubActivityForm.tsx
@@ -1,12 +1,11 @@
 import React, { FunctionComponent } from 'react';
 import { Table } from 'bluesquare-components';
-import { Box, Grid } from '@mui/material';
+import { Box } from '@mui/material';
 import { Round } from '../../../constants/types';
 import { useTableState } from '../../../../../../../hat/assets/js/apps/Iaso/utils/table';
 import { useGetSubActivities } from './hooks/api/useGetSubActivities';
 import { useSubActivitiesColumns } from './hooks/useSubActivitiesColumns';
-import { CreateSubActivity } from './components/Modal/CreateEditSubActivity';
-import { RoundDates } from './components/RoundDates';
+import { SubActivitiesInfos } from './components/SubActivitiesInfos';
 
 type Props = { round?: Round };
 
@@ -20,15 +19,8 @@ export const SubActivityForm: FunctionComponent<Props> = ({ round }) => {
     const columns = useSubActivitiesColumns(round);
     return (
         <>
-            <Box ml={2} mt={4} mb={-4}>
-                <Grid container spacing={2}>
-                    <Grid item xs={8}>
-                        <RoundDates round={round} />
-                    </Grid>
-                    <Grid container item xs={4} justifyContent="flex-end">
-                        <CreateSubActivity iconProps={{}} round={round} />
-                    </Grid>
-                </Grid>
+            <Box mb={-6}>
+                <SubActivitiesInfos round={round} />
             </Box>
             <Box
                 sx={{
